str_count {stringr}	R Documentation

Count the number of matches in a string.

Description

Vectorised over string and pattern, shorter is recycled to same length as longest.

Usage

  str_count(string, pattern)

Arguments

string
input vector. This must be an atomic vector, and will be coerced to a character vector

pattern
pattern to look for, as defined by a POSIX regular expression. See the “Extended Regular Expressions” section of regex for details. See fixed, ignore.case and perl for how to use other types of matching: fixed, case insensitive and perl-compatible.

Value

integer vector

See Also

regexpr which this function wraps

str_locate/str_locate_all to locate position of matches

Examples

fruit <- c("apple", "banana", "pear", "pineapple")
str_count(fruit, "a")
str_count(fruit, "p")
str_count(fruit, "e")
str_count(fruit, c("a", "b", "p", "p"))

[Package stringr version 0.6.2 Index]