str_detect {stringr}	R Documentation

Detect the presence or absence of a pattern in a string.

Description

Vectorised over string and pattern.

Usage

  str_detect(string, pattern)

Arguments

string
input vector. This must be an atomic vector, and will be coerced to a character vector

pattern
pattern to look for, as defined by a POSIX regular expression. See the “Extended Regular Expressions” section of regex for details. See fixed, ignore.case and perl for how to use other types of matching: fixed, case insensitive and perl-compatible.

Value

boolean vector

See Also

grepl which this function wraps

Examples

fruit <- c("apple", "banana", "pear", "pinapple")
str_detect(fruit, "a")
str_detect(fruit, "^a")
str_detect(fruit, "a$")
str_detect(fruit, "b")
str_detect(fruit, "[aeiou]")

# Also vectorised over pattern
str_detect("aecfg", letters)

[Package stringr version 0.6.2 Index]