str_extract_all {stringr}	R Documentation

Extract all pieces of a string that match a pattern.

Description

Vectorised over string. pattern should be a single pattern, i.e. a character vector of length one.

Usage

  str_extract_all(string, pattern)

Arguments

string
input vector. This must be an atomic vector, and will be coerced to a character vector

pattern
pattern to look for, as defined by a POSIX regular expression. See the “Extended Regular Expressions” section of regex for details. See fixed, ignore.case and perl for how to use other types of matching: fixed, case insensitive and perl-compatible.

Value

list of character vectors.

See Also

str_extract to extract the first match

Examples

shopping_list <- c("apples x4", "bag of flour", "bag of sugar", "milk x2")
str_extract_all(shopping_list, "[a-z]+")
str_extract_all(shopping_list, "\\b[a-z]+\\b")
str_extract_all(shopping_list, "\\d")

[Package stringr version 0.6.2 Index]