str_length {stringr}	R Documentation

The length of a string (in characters).

Description

The length of a string (in characters).

Usage

  str_length(string)

Arguments

string
input vector. This must be an atomic vector, and will be coerced to a character vector

Value

numeric vector giving number of characters in each element of the character vector. Missing string have missing length.

See Also

nchar which this function wraps

Examples

str_length(letters)
str_length(c("i", "like", "programming", NA))

[Package stringr version 0.6.2 Index]