str_match_all {stringr}	R Documentation

Extract all matched groups from a string.

Description

Vectorised over string. pattern should be a single pattern, i.e. a character vector of length one.

Usage

  str_match_all(string, pattern)

Arguments

pattern
pattern to look for, as defined by a POSIX regular expression. Pattern should contain groups, defined by (). See the “Extended Regular Expressions” section of regex for details.

string
input vector. This must be an atomic vector, and will be coerced to a character vector

Value

list of character matrices, as given by str_match

Examples

strings <- c("Home: 219 733 8965.  Work: 229-293-8753 ",
  "banana pear apple", "595 794 7569 / 387 287 6718")
phone <- "([2-9][0-9]{2})[- .]([0-9]{3})[- .]([0-9]{4})"

str_extract_all(strings, phone)
str_match_all(strings, phone)

[Package stringr version 0.6.2 Index]