Replace matched patterns in a string.

Vectorised over string, pattern and replacement.

str_replace(string, pattern, replacement)

str_replace_all(string, pattern, replacement)

Arguments

string

Input vector. Either a character vector, or something coercible to one.

pattern, replacement

Supply separate pattern and replacement strings to vectorise over the patterns. References of the form \1, \2 will be replaced with the contents of the respective matched group (created by ()) within the pattern. For str_replace_all only, you can perform multiple patterns and replacements to each string, by passing a named character to pattern. Alternatively, you can supply use a replacement function: it will be called once for each match and its return value will be used to replace the match.

Value

A character vector.

See also

str_replace_na to turn missing values into "NA"; stri_replace for the underlying implementation.

Examples

fruits <- c("one apple", "two pears", "three bananas")
str_replace(fruits, "[aeiou]", "-")

#> [1] "-ne apple"     "tw- pears"     "thr-e bananas"
#> 

str_replace_all(fruits, "[aeiou]", "-")

#> [1] "-n- -ppl-"     "tw- p--rs"     "thr-- b-n-n-s"
#> 

str_replace_all(fruits, "[aeiou]", toupper)

#> [1] "OnE ApplE"     "twO pEArs"     "thrEE bAnAnAs"
#> 

str_replace(fruits, "([aeiou])", "")

#> [1] "ne apple"     "tw pears"     "thre bananas"
#> 

str_replace(fruits, "([aeiou])", "\\1\\1")

#> [1] "oone apple"     "twoo pears"     "threee bananas"
#> 

str_replace(fruits, "[aeiou]", c("1", "2", "3"))

#> [1] "1ne apple"     "tw2 pears"     "thr3e bananas"
#> 

str_replace(fruits, c("a", "e", "i"), "-")

#> [1] "one -pple"     "two p-ars"     "three bananas"
#> 

fruits <- c("one apple", "two pears", "three bananas")
str_replace(fruits, "[aeiou]", "-")

#> [1] "-ne apple"     "tw- pears"     "thr-e bananas"
#> 

str_replace_all(fruits, "[aeiou]", "-")

#> [1] "-n- -ppl-"     "tw- p--rs"     "thr-- b-n-n-s"
#> 

str_replace_all(fruits, "([aeiou])", "")

#> [1] "n ppl"    "tw prs"   "thr bnns"
#> 

str_replace_all(fruits, "([aeiou])", "\\1\\1")

#> [1] "oonee aapplee"      "twoo peeaars"       "threeee baanaanaas"
#> 

str_replace_all(fruits, "[aeiou]", c("1", "2", "3"))

#> [1] "1n1 1ppl1"     "tw2 p22rs"     "thr33 b3n3n3s"
#> 

str_replace_all(fruits, c("a", "e", "i"), "-")

#> [1] "one -pple"     "two p-ars"     "three bananas"
#> 

# If you want to apply multiple patterns and replacements to the same
# string, pass a named version to pattern.
fruits %>%
  str_c(collapse = "---") %>%
  str_replace_all(c("one" = 1, "two" = 2, "three" = 3))

#> [1] "1 apple---2 pears---3 bananas"
#> 

# Use a function for more sophisticated replacement. This example
# replaces colour names with their hex values.
colours <- str_c("\\b", colors(), "\\b", collapse="|")
col2hex <- function(col) {
  rgb <- col2rgb(col)
  rgb(rgb["red", ], rgb["green", ], rgb["blue", ], max = 255)
}

x <- c(
  "Roses are red, violets are blue",
  "My favourite colour is green"
)
str_replace_all(x, colours, col2hex)

#> [1] "Roses are #FF0000, violets are #0000FF"
#> [2] "My favourite colour is #00FF00"        
#>