Detect the presence or absence of a pattern at the beginning or end of a string.

Vectorised over string and pattern.

str_starts(string, pattern, negate = FALSE)

str_ends(string, pattern, negate = FALSE)

Arguments

string
Input vector. Either a character vector, or something coercible to one.

pattern
Pattern with which the string starts or ends.

The default interpretation is a regular expression, as described in stringi::stringi-search-regex. Control options with regex().

Match a fixed string (i.e. by comparing only bytes), using fixed(). This is fast, but approximate. Generally, for matching human text, you'll want coll() which respects character matching rules for the specified locale.

negate
If TRUE, return non-matching elements.

Value

A logical vector.

See also

Examples

fruit <- c("apple", "banana", "pear", "pineapple")
str_starts(fruit, "p")

#> [1] FALSE FALSE  TRUE  TRUE

str_starts(fruit, "p", negate = TRUE)

#> [1]  TRUE  TRUE FALSE FALSE

str_ends(fruit, "e")

#> [1]  TRUE FALSE FALSE  TRUE

str_ends(fruit, "e", negate = TRUE)

#> [1] FALSE  TRUE  TRUE FALSE