Keep strings matching a pattern.

This is a convenient wrapper around x[str_detect(x, pattern)]. Vectorised over string and pattern

str_subset(string, pattern)

Arguments

string

Input vector. Either a character vector, or something coercible to one.

pattern

Pattern to look for. The default interpretation is a regular expression, as described in stringi-search-regex. Control options with regex(). Match a fixed string (i.e. by comparing only bytes), using fixed(x). This is fast, but approximate. Generally, for matching human text, you'll want coll(x) which respects character matching rules for the specified locale. Match character, word, line and sentence boundaries with boundary(). An empty pattern, "", is equivalent to boundary("character").

Value

A character vector.

See also

grep with argument value = TRUE, stri_subset for the underlying implementation.

Examples

fruit <- c("apple", "banana", "pear", "pinapple")
str_subset(fruit, "a")

#> [1] "apple"    "banana"   "pear"     "pinapple"
#> 

str_subset(fruit, "^a")

#> [1] "apple"
#> 

str_subset(fruit, "a$")

#> [1] "banana"
#> 

str_subset(fruit, "b")

#> [1] "banana"
#> 

str_subset(fruit, "[aeiou]")

#> [1] "apple"    "banana"   "pear"     "pinapple"
#> 

# Missings are silently dropped
str_subset(c("a", NA, "b"), ".")

#> [1] "a" "b"
#>