str_escape {stringr}	R Documentation

Escape regular expression metacharacters

Description

This function escapes metacharacter, the characters that have special meaning to the regular expression engine. In most cases you are better off using fixed() since it is faster, but str_escape() is useful if you are composing user provided strings into a pattern.

Usage


str_escape(string)

Arguments

string
Input vector. Either a character vector, or something coercible to one.

Value

A character vector the same length as string.

Examples


str_detect(c("a", "."), ".")
str_detect(c("a", "."), str_escape("."))

[Package stringr version 1.5.1 Index]