str_subset {stringr}	R Documentation

Keep strings matching a pattern.

Description

This is a convenient wrapper around x[str_detect(x, pattern)]. Vectorised over string and pattern

Usage

str_subset(string, pattern)

Arguments

string
Input vector. Either a character vector, or something coercible to one.

pattern
Pattern to look for.

The default interpretation is a regular expression, as described in stringi-search-regex. Control options with regex().

Match a fixed string (i.e. by comparing only bytes), using fixed(x). This is fast, but approximate. Generally, for matching human text, you'll want coll(x) which respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with boundary(). An empty pattern, "", is equivalent to boundary("character").

Value

A character vector.

See Also

grep with argument value = TRUE, stri_subset for the underlying implementation.

Examples

fruit <- c("apple", "banana", "pear", "pinapple")
str_subset(fruit, "a")
str_subset(fruit, "^a")
str_subset(fruit, "a$")
str_subset(fruit, "b")
str_subset(fruit, "[aeiou]")

# Missings are silently dropped
str_subset(c("a", NA, "b"), ".")

[Package stringr version 1.0.0 Index]