Keep strings matching a pattern, or find positions.

str_subset() is a wrapper around x[str_detect(x, pattern)], and is equivalent to grep(pattern, x, value = TRUE). str_which() is a wrapper around which(str_detect(x, pattern)), and is equivalent to grep(pattern, x). See str_detect() for an equivalent to grepl(pattern, x).

str_subset(string, pattern, negate = FALSE)

str_which(string, pattern, negate = FALSE)

Arguments

string
Input vector. Either a character vector, or something coercible to one.

pattern
Pattern to look for.

The default interpretation is a regular expression, as described in stringi::stringi-search-regex. Control options with regex().

Match a fixed string (i.e. by comparing only bytes), using fixed(). This is fast, but approximate. Generally, for matching human text, you'll want coll() which respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with boundary(). An empty pattern, "", is equivalent to boundary("character").

negate
If TRUE, return non-matching elements.

Value

A character vector.

Details

Vectorised over string and pattern

See also

Examples

fruit <- c("apple", "banana", "pear", "pinapple")
str_subset(fruit, "a")

#> [1] "apple"    "banana"   "pear"     "pinapple"

str_which(fruit, "a")

#> [1] 1 2 3 4

str_subset(fruit, "^a")

#> [1] "apple"

str_subset(fruit, "a$")

#> [1] "banana"

str_subset(fruit, "b")

#> [1] "banana"

str_subset(fruit, "[aeiou]")

#> [1] "apple"    "banana"   "pear"     "pinapple"

# Returns elements that do NOT match
str_subset(fruit, "^p", negate = TRUE)

#> [1] "apple"  "banana"

# Missings never match
str_subset(c("a", NA, "b"), ".")

#> [1] "a" "b"

str_which(c("a", NA, "b"), ".")

#> [1] 1 3