stringr-package {stringr}	R Documentation

stringr: Simple, Consistent Wrappers for Common String Operations

Description

A consistent, simple and easy to use set of wrappers around the fantastic 'stringi' package. All function and argument names (and positions) are consistent, all functions deal with "NA"'s and zero length vectors in the same way, and the output from one function is easy to feed into the input of another.

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com [copyright holder]

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://stringr.tidyverse.org

https://github.com/tidyverse/stringr

Report bugs at https://github.com/tidyverse/stringr/issues

[Package stringr version 1.5.0 Index]