Extract words from a sentence.

Extract words from a sentence.

word(string, start = 1L, end = start, sep = fixed(" "))

Arguments

string
input character vector.

start
integer vector giving position of first word to extract. Defaults to first word. If negative, counts backwards from last character.

end
integer vector giving position of last word to extract. Defaults to first word. If negative, counts backwards from last character.

sep
separator between words. Defaults to single space.

Value

character vector of words from start to end (inclusive). Will be length of longest input argument.

Examples

sentences <- c("Jane saw a cat", "Jane sat down")
word(sentences, 1)

#> [1] "Jane" "Jane"

word(sentences, 2)

#> [1] "saw" "sat"

word(sentences, -1)

#> [1] "cat"  "down"

word(sentences, 2, -1)

#> [1] "saw a cat" "sat down" 

# Also vectorised over start and end
word(sentences[1], 1:3, -1)

#> [1] "Jane saw a cat" "saw a cat"      "a cat"         

word(sentences[1], 1, 1:4)

#> [1] "Jane"           "Jane saw"       "Jane saw a"     "Jane saw a cat"

# Can define words by other separators
str <- 'abc.def..123.4568.999'
word(str, 1, sep = fixed('..'))

#> [1] "abc.def"

word(str, 2, sep = fixed('..'))

#> [1] "123.4568.999"